Move protobuf output to /pkg dir
